requirements.txt
vizdoom
numpy
Pillow
requests
awscli
boto3
nats-py